requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2==2.9.10
channels==4.1.0
channels-redis==4.2.0
redis==5.0.8 